Update based on PR comments
