+++ backend/spec/controllers/spree/admin/orders_controller_spec.rb
@@ -350,55 +350,4 @@ describe Spree::Admin::OrdersController do
     end
 
   end
-
-  describe "#short_ship_units" do
-    subject { spree_post :short_ship_units, id: order.number, inventory_unit_ids: inventory_units.map(&:id) }
-
-    let(:order) { order = create(:order_ready_to_ship, :number => "R100", :state => "complete") }
-    let(:referer) { "order_admin_page" }
-
-    before { request.env["HTTP_REFERER"] = referer }
-
-    context "no inventory unit ids are provided" do
-      let(:inventory_units) { [] }
-
-      it "redirects back" do
-        subject
-        response.should redirect_to referer
-      end
-
-      it "sets an error message" do
-        subject
-        expect(flash[:error]).to eq Spree.t(:no_inventory_selected)
-      end
-    end
-
-    context "unable to find all the inventory" do
-      let(:inventory_units) { [Spree::InventoryUnit.new(id: Spree::InventoryUnit.last.id + 1)] }
-
-      it "redirects back" do
-        subject
-        response.should redirect_to referer
-      end
-
-      it "sets an error message" do
-        subject
-        expect(flash[:error]).to eq Spree.t(:unable_to_find_all_inventory_units)
-      end
-    end
-
-    context "successfully cancels inventory" do
-      let(:inventory_units) { order.inventory_units.not_canceled }
-
-      it "redirects to admin order edit" do
-        subject
-        response.should redirect_to(spree.edit_admin_order_path(order))
-      end
-
-      it "sets an success message" do
-        subject
-        expect(flash[:success]).to eq Spree.t(:inventory_canceled)
-      end
-    end
-  end
 end
